add pep, remove read_json_file file
