+++ lab_2/main.py
@@ -1,11 +1,28 @@
-import math
 import json
 
 from scipy.special import erfc
 from mpmath import gammainc
 from math import sqrt, factorial
-from work_with_json_file import read_json_file
 from constants import BLOCK_SIZE, PI_VALUES, PATH
+
+
+def read_json_file(file_path: str) -> dict:
+    """
+    a method for reading paths from a json file
+    parametrs: file_path as str
+    return: dict
+
+    """
+    try:
+        with open(file_path, "r", encoding="utf-8") as file:
+            data = json.load(file)
+            return data
+    except FileNotFoundError:
+        print("Файл не найден.")
+    except json.JSONDecodeError:
+        print("Ошибка при декодировании JSON-данных.")
+    except Exception as e:
+        print(f"Произошла ошибка: {e}")
 
 
 def frequency_bit_test(sequence: str) -> float:
@@ -85,7 +102,7 @@ def long_sequence_units_test(sequence: str) -> float:
         return None
 
 
-def main():
+def main() -> None:
     data = read_json_file(PATH)
     for language, sequence in data.items():
         p_value_frequency_bit_test = frequency_bit_test(sequence)
